+++ app/models/data_import.rb
@@ -346,6 +346,15 @@ class DataImport < Sequel::Model
 
   private
 
+  def get_provider_name_from_id(service_item_id)
+    begin
+      connector_params = JSON.parse(service_item_id)
+      return connector_params['provider']
+    rescue StandardError
+      return nil
+    end
+  end
+
   def dispatch
     self.state = STATE_UPLOADING
     return from_table         if table_copy.present? || from_query.present?
@@ -674,7 +683,8 @@ class DataImport < Sequel::Model
   # * importer: the new importer (nil if download errors detected)
   # * connector: the connector that the importer uses
   def new_importer_with_connector
-    CartoDB::Importer2::ConnectorRunner.check_availability!(current_user)
+    provider_name = get_provider_name_from_id(service_item_id)
+    CartoDB::Importer2::ConnectorRunner.check_availability!(current_user, provider_name)
 
     database_options = pg_options